refactor(infra): unmask api_jobs dead-code, disambiguate db/deps module names
Remove the blanket `#![allow(dead_code)]` covering the whole ~3000-LOC
api_jobs remote-job subsystem. The only item it was hiding is
`JobStore::create`, a snapshot-less convenience used solely by the store's
unit tests; gate it behind `#[cfg(test)]` so production builds carry no dead
weight and the lint stays honest. All other api_jobs items are live.

Rename `core/db/tunnel.rs` -> `core/db/ssh_forward.rs`: it's an ad-hoc SSH DB
port-forward (default 33306) and the old name collided with the unrelated
`core/tunnel` ServiceTunnel entity. Module + re-export wiring only;
public symbols and behavior unchanged.

Rename `core/deps/stack.rs` -> `core/deps/dependency_graph.rs`: it's a
dependency-edge graph and the old name collided with the unrelated
`core/stack` combined-fixes feature. Module + re-export wiring only;
public symbols and behavior unchanged.

Refs #6775.
