+++ src/core/api_jobs/mod.rs
@@ -1,5 +1,3 @@
-#![allow(dead_code)]
-
 mod persistence;
 mod remote_runner;
 mod store;
