+++ src/core/deps/dependency_graph.rs
